+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+## [v2.5.1-alpha] 
+
+Adding contents type conversions to forge inputs.
+
+
+## [v2.5.0-alpha] 
+
+Adding Forge functions. 
 
 ## [v2.4.4-alpha]